backend/requirements.txt
Flask>=2.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-jwt-extended>=4.4.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
bcrypt>=4.0.0